website/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
googletrans==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.4.1
pytesseract==0.2.6
requests==2.21.0
urllib3==1.24.1
Werkzeug==0.14.1
xmltodict==0.12.0
flask-cors
